Fix spell check false positive by ignoring word (#38)
In the latest release of the codespell tool used for automated spell checking of the files of this project, the word "tabl" was added to the codespell misspelled words dictionary as a misspelling of "table".

This work occurs in a string in the library's source code, which causes the spell check to fail. Although it is not clear to me what the purpose of this string is, I will assume it is correct and intended as it is. For this reason, the false positive is resolved by configuring codespell to ignore the problematic word.