Adjust spell checker configuration to fix false positive
The codespell tool is used for automated detection of common misspellings in the files of this project.

A new release of codespell includes an expansion of its misspellings dictionary. One of the additions is "secur" as a
misspelling of "secure". The word "secur" occurs in the library's codebase. Although it is not clear why the author felt
the need to shorten the words, the prevalence of such shortening clearly indicate it was done on purpose so this is a
false positive.

The false positive is fixed by adding "secur" to the ignored words list in the codespell configuration file.