Fix spell check false positive by ignoring word
The codespell spellchecker tool is used to automatically detect commonly misspelled words in the files of this project.

The misspelled words dictionary was expanded in the latest release of codespell. Some of the text in the project
codebase happens to match against newly added entries, which caused codespell to produce a false misspelled word
detection.

Since the code that produced the detection is correct and intended, the false positive is resolved by configuring
codespell to ignore the problematic word.